+++ ftag/mock.py
@@ -106,11 +106,11 @@ def get_mock_scores(labels: np.ndarray, is_xbb: bool = False) -> np.ndarray:
     for i in range(n_classes):
         tmp_means = []
         tmp_means = [
-            0 if j != i else mean_scale_list[np.random.randint(0, len(mean_scale_list))]
+            0 if j != i else mean_scale_list[rng.integers(0, len(mean_scale_list))]
             for j in range(n_classes)
         ]
         means.append(tmp_means)
-        scales.append(mean_scale_list[np.random.randint(0, len(mean_scale_list))])
+        scales.append(mean_scale_list[rng.integers(0, len(mean_scale_list))])
 
     # Map the labels to the means
     label_mapping = dict(zip(label_dict.values(), means))